other_games/requirements_python37.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astroid=2.4.2=pypi_0
atari-py=0.2.6=pypi_0
blas=1.0=mkl
box2d-py=2.3.8=pypi_0
ca-certificates=2020.7.22=0
certifi=2020.6.20=py37_0
cloudpickle=1.3.0=pypi_0
cpuonly=1.0=0
cycler=0.10.0=pypi_0
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=pypi_0
gym=0.17.3=pypi_0
gym-retro=0.8.0=pypi_0
intel-openmp=2020.2=254
isort=5.5.3=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lazy-object-proxy=1.4.3=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=he6710b0_1
matplotlib=3.3.2=pypi_0
mccabe=0.6.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.1=py37hfd86e86_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
opencv-python=4.4.0.44=pypi_0
openssl=1.1.1h=h7b6447c_0
pandas=1.1.4=pypi_0
pillow=7.2.0=py37hb39fc2d_0
pip=20.2.2=py37_0
pyglet=1.5.0=pypi_0
pylint=2.6.0=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pypi_0
pytorch=1.4.0=py3.7_cpu_0
pytz=2020.4=pypi_0
readline=8.0=h7b6447c_0
scipy=1.5.2=pypi_0
seaborn=0.11.0=pypi_0
setuptools=49.6.0=py37_1
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
toml=0.10.1=pypi_0
torchvision=0.5.0=py37_cpu
typed-ast=1.4.1=pypi_0
wheel=0.35.1=py_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
